+++ tcod/__init__.py
@@ -17,5 +17,15 @@
 """
 from __future__ import absolute_import
 
+import sys
+
+import warnings
+
 from tcod.libtcodpy import *
 from tcod.version import __version__
+
+if sys.version_info[0] == 2:
+    warnings.warn(
+        "python-tcod will drop support for Python 2.7 in the near future.",
+        DeprecationWarning
+    )